requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.2
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
decorator==4.4.2
distlib==0.3.0
filelock==3.0.12
future==0.18.2
identify==1.4.19
idna==2.9
importlib-metadata==1.6.1
more-itertools==8.4.0
nodeenv==1.4.0
packaging==20.4
pluggy==0.13.1
pre-commit==2.5.1
py==1.8.2
pycparser==2.20
pyparsing==2.4.7
pytest==5.4.3
python-dotenv==0.13.0
PyYAML==5.3.1
requests==2.23.0
requests-mock==1.8.0
six==1.15.0
toml==0.10.1
tornado==6.0.4
urllib3==1.25.9
virtualenv==20.0.23
wcwidth==0.2.4
zipp==3.1.0
